added cycle controller for repeat,reverse,forever features

+++ src/Types.h
@@ -2,6 +2,8 @@
 #ifndef TYPES_H
 #define TYPES_H
 
-typedef unsigned long Uint32;
+#include <stdint.h>
+
+typedef uint32_t Uint32;
 
 #endif
